UV/Vis Python Automation Script
UV/Vis python automation script to process UV/Vis csv data and plot the result. It will also ask for prompts to calculate the band gap from inputted onset of absorbance from intersection of line inputs. It will output a picture of the graph, calculations, and an excel file with values if user wants to plot in excel as well.